// frontend/src/api/api-client/models/UncertainBoolean.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ISI Backend API
 * ISI Backend - Serivce für das Informationssystem für soziale Infrastrukturplanung
 *
 * The version of the OpenAPI document: 0.0.1-SNAPSHOT
 * Contact: noreply@mail.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const UncertainBoolean = {
    Unspecified: 'UNSPECIFIED',
    True: 'TRUE',
    False: 'FALSE'
} as const;
export type UncertainBoolean = typeof UncertainBoolean[keyof typeof UncertainBoolean];


export function UncertainBooleanFromJSON(json: any): UncertainBoolean {
    return UncertainBooleanFromJSONTyped(json, false);
}

export function UncertainBooleanFromJSONTyped(json: any, ignoreDiscriminator: boolean): UncertainBoolean {
    return json as UncertainBoolean;
}

export function UncertainBooleanToJSON(value?: UncertainBoolean | null): any {
    return value as any;
}